Poly tool now supports single click clones but with slight behaviour change.
Poly tool needed a slight behaviour tool if it was to support single click
duplicates of the previous shape created (just like all other shape tools)
as the behaviour previoulsy allowed a single click to place the first point.
Poly tool now needs a hold and drag operation to start it off. Not 100% in
favour of this but it is now consistant with other tools and has the ability
to duplicate.