Update README.md
Fill in missing sections using Browse and Chat activities as a template.

Describe dependencies.

Demote the Box2D instructions as being for development purposes mainly; as Box2D is available from Debian and Fedora distributions.

Remove unmaintained dfsg branch.  It was to follow along with the Debian Free Software Guidelines inspired trimming of the source to remove binaries.  As we have now removed binaries, it's the same as master.

Co-authored-by: James Cameron <quozl@laptop.org>