Close the create_toolbar by default
sugargame doesn't detect resize due to open/closing of menu by `create_toolbar` (and other similar dropdown widgets). This leads to a horizontal gap because the canvas is not resized when the menu is closed.

Also, using the color selector dropdown does trigger the resize and the pygame canvas fills the gap. But now all the elements are shifted up because (0, 0) of pygame canvas is shifted up.

A solution to all of this would be to not show the `create_toolbar` by default, eliminating the need to resize the canvas anytime.